Saving progress as I didn't save since a long time
- Manual API is supported by `Thoth.Json.Core`, `Thoth.Json.JavaScript`, `Thoth.Json.Newtonsoft`
- Tests structure is slowly taking shape
