+++ Directory.Packages.props
@@ -3,9 +3,16 @@
     <ManagePackageVersionsCentrally>true</ManagePackageVersionsCentrally>
   </PropertyGroup>
   <ItemGroup>
-    <PackageVersion Update="FSharp.Core" Version="4.7.2" />
+    <PackageVersion Include="BlackFox.CommandLine" Version="1.0.0" />
+    <PackageVersion Include="FSharp.Core" Version="4.7.2" />
+    <!--
+        We use an older version of Expecto to keep FSharp.Core version as low as possible
+    -->
+    <PackageVersion Include="Expecto" Version="9.0.4" />
     <PackageVersion Include="Fable.Core" Version="3.6.2" />
     <PackageVersion Include="Fable.Mocha" Version="2.17.0" />
-    <PackageVersion Include="Newtonsoft.Json" Version="11.0.2" />
+    <PackageVersion Include="Newtonsoft.Json" Version="13.0.1" />
+    <PackageVersion Include="Semver" Version="2.3.0" />
+    <PackageVersion Include="SimpleExec" Version="11.0.0" />
   </ItemGroup>
 </Project>
