+++ Thoth.Json.sln
@@ -3,8 +3,6 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
 VisualStudioVersion = 15.0.26124.0
 MinimumVisualStudioVersion = 15.0.26124.0
-Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Thoth.Json", "src\Thoth.Json.fsproj", "{E4905A84-E416-47B4-8BFB-BB02446B74EE}"
-EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "packages", "packages", "{D009BA72-08B7-4F6A-987D-806FFB01AE88}"
 EndProject
 Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Thoth.Json.Core", "packages\Thoth.Json.Core\Thoth.Json.Core.fsproj", "{89A13A89-F215-4A35-B5BD-CAB48BAAF72F}"
@@ -19,7 +17,13 @@ Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Prototype.JavaScript", "pac
 EndProject
 Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Prototype.DotNet", "packages\Prototype.DotNet\Prototype.DotNet.fsproj", "{BF33FC29-2FBD-4DBC-97D7-9CF207130908}"
 EndProject
-Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Thoth.Json.JavaScript.Tests", "tests\Thoth.Json.JavaScript.Tests.fsproj", "{34DED59E-B475-41B5-9633-EE580C967BAB}"
+Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Build", "build\Build.fsproj", "{C92D7510-F8E4-491F-9508-6B8BF7FF0408}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{BC3238CA-1D20-4D33-B633-033A2627B1DF}"
+EndProject
+Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Thoth.Json.JavaScript.Tests", "tests\Thoth.Json.Tests.JavaScript\Thoth.Json.Tests.JavaScript.fsproj", "{82CAAF28-9AE0-4515-A47E-E436D0D3FA05}"
+EndProject
+Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Thoth.Json.Newtonsoft.Tests", "tests\Thoth.Json.Tests.Newtonsoft\Thoth.Json.Tests.Newtonsoft.fsproj", "{DB50A83E-B1BA-44C1-BC55-760DCA114B2F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -34,18 +38,6 @@ Global
 		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{E4905A84-E416-47B4-8BFB-BB02446B74EE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E4905A84-E416-47B4-8BFB-BB02446B74EE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E4905A84-E416-47B4-8BFB-BB02446B74EE}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{E4905A84-E416-47B4-8BFB-BB02446B74EE}.Debug|x64.Build.0 = Debug|Any CPU
-		{E4905A84-E416-47B4-8BFB-BB02446B74EE}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{E4905A84-E416-47B4-8BFB-BB02446B74EE}.Debug|x86.Build.0 = Debug|Any CPU
-		{E4905A84-E416-47B4-8BFB-BB02446B74EE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E4905A84-E416-47B4-8BFB-BB02446B74EE}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E4905A84-E416-47B4-8BFB-BB02446B74EE}.Release|x64.ActiveCfg = Release|Any CPU
-		{E4905A84-E416-47B4-8BFB-BB02446B74EE}.Release|x64.Build.0 = Release|Any CPU
-		{E4905A84-E416-47B4-8BFB-BB02446B74EE}.Release|x86.ActiveCfg = Release|Any CPU
-		{E4905A84-E416-47B4-8BFB-BB02446B74EE}.Release|x86.Build.0 = Release|Any CPU
 		{89A13A89-F215-4A35-B5BD-CAB48BAAF72F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{89A13A89-F215-4A35-B5BD-CAB48BAAF72F}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{89A13A89-F215-4A35-B5BD-CAB48BAAF72F}.Debug|x64.ActiveCfg = Debug|Any CPU
@@ -118,18 +110,42 @@ Global
 		{BF33FC29-2FBD-4DBC-97D7-9CF207130908}.Release|x64.Build.0 = Release|Any CPU
 		{BF33FC29-2FBD-4DBC-97D7-9CF207130908}.Release|x86.ActiveCfg = Release|Any CPU
 		{BF33FC29-2FBD-4DBC-97D7-9CF207130908}.Release|x86.Build.0 = Release|Any CPU
-		{34DED59E-B475-41B5-9633-EE580C967BAB}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{34DED59E-B475-41B5-9633-EE580C967BAB}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{34DED59E-B475-41B5-9633-EE580C967BAB}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{34DED59E-B475-41B5-9633-EE580C967BAB}.Debug|x64.Build.0 = Debug|Any CPU
-		{34DED59E-B475-41B5-9633-EE580C967BAB}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{34DED59E-B475-41B5-9633-EE580C967BAB}.Debug|x86.Build.0 = Debug|Any CPU
-		{34DED59E-B475-41B5-9633-EE580C967BAB}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{34DED59E-B475-41B5-9633-EE580C967BAB}.Release|Any CPU.Build.0 = Release|Any CPU
-		{34DED59E-B475-41B5-9633-EE580C967BAB}.Release|x64.ActiveCfg = Release|Any CPU
-		{34DED59E-B475-41B5-9633-EE580C967BAB}.Release|x64.Build.0 = Release|Any CPU
-		{34DED59E-B475-41B5-9633-EE580C967BAB}.Release|x86.ActiveCfg = Release|Any CPU
-		{34DED59E-B475-41B5-9633-EE580C967BAB}.Release|x86.Build.0 = Release|Any CPU
+		{C92D7510-F8E4-491F-9508-6B8BF7FF0408}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C92D7510-F8E4-491F-9508-6B8BF7FF0408}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C92D7510-F8E4-491F-9508-6B8BF7FF0408}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{C92D7510-F8E4-491F-9508-6B8BF7FF0408}.Debug|x64.Build.0 = Debug|Any CPU
+		{C92D7510-F8E4-491F-9508-6B8BF7FF0408}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{C92D7510-F8E4-491F-9508-6B8BF7FF0408}.Debug|x86.Build.0 = Debug|Any CPU
+		{C92D7510-F8E4-491F-9508-6B8BF7FF0408}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C92D7510-F8E4-491F-9508-6B8BF7FF0408}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C92D7510-F8E4-491F-9508-6B8BF7FF0408}.Release|x64.ActiveCfg = Release|Any CPU
+		{C92D7510-F8E4-491F-9508-6B8BF7FF0408}.Release|x64.Build.0 = Release|Any CPU
+		{C92D7510-F8E4-491F-9508-6B8BF7FF0408}.Release|x86.ActiveCfg = Release|Any CPU
+		{C92D7510-F8E4-491F-9508-6B8BF7FF0408}.Release|x86.Build.0 = Release|Any CPU
+		{82CAAF28-9AE0-4515-A47E-E436D0D3FA05}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{82CAAF28-9AE0-4515-A47E-E436D0D3FA05}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{82CAAF28-9AE0-4515-A47E-E436D0D3FA05}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{82CAAF28-9AE0-4515-A47E-E436D0D3FA05}.Debug|x64.Build.0 = Debug|Any CPU
+		{82CAAF28-9AE0-4515-A47E-E436D0D3FA05}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{82CAAF28-9AE0-4515-A47E-E436D0D3FA05}.Debug|x86.Build.0 = Debug|Any CPU
+		{82CAAF28-9AE0-4515-A47E-E436D0D3FA05}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{82CAAF28-9AE0-4515-A47E-E436D0D3FA05}.Release|Any CPU.Build.0 = Release|Any CPU
+		{82CAAF28-9AE0-4515-A47E-E436D0D3FA05}.Release|x64.ActiveCfg = Release|Any CPU
+		{82CAAF28-9AE0-4515-A47E-E436D0D3FA05}.Release|x64.Build.0 = Release|Any CPU
+		{82CAAF28-9AE0-4515-A47E-E436D0D3FA05}.Release|x86.ActiveCfg = Release|Any CPU
+		{82CAAF28-9AE0-4515-A47E-E436D0D3FA05}.Release|x86.Build.0 = Release|Any CPU
+		{DB50A83E-B1BA-44C1-BC55-760DCA114B2F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{DB50A83E-B1BA-44C1-BC55-760DCA114B2F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{DB50A83E-B1BA-44C1-BC55-760DCA114B2F}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{DB50A83E-B1BA-44C1-BC55-760DCA114B2F}.Debug|x64.Build.0 = Debug|Any CPU
+		{DB50A83E-B1BA-44C1-BC55-760DCA114B2F}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{DB50A83E-B1BA-44C1-BC55-760DCA114B2F}.Debug|x86.Build.0 = Debug|Any CPU
+		{DB50A83E-B1BA-44C1-BC55-760DCA114B2F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{DB50A83E-B1BA-44C1-BC55-760DCA114B2F}.Release|Any CPU.Build.0 = Release|Any CPU
+		{DB50A83E-B1BA-44C1-BC55-760DCA114B2F}.Release|x64.ActiveCfg = Release|Any CPU
+		{DB50A83E-B1BA-44C1-BC55-760DCA114B2F}.Release|x64.Build.0 = Release|Any CPU
+		{DB50A83E-B1BA-44C1-BC55-760DCA114B2F}.Release|x86.ActiveCfg = Release|Any CPU
+		{DB50A83E-B1BA-44C1-BC55-760DCA114B2F}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{89A13A89-F215-4A35-B5BD-CAB48BAAF72F} = {D009BA72-08B7-4F6A-987D-806FFB01AE88}
@@ -138,5 +154,7 @@ Global
 		{0FDDC376-8B73-43C8-B0B1-4665CAAF5238} = {D009BA72-08B7-4F6A-987D-806FFB01AE88}
 		{D04F052E-C557-4F61-8FAB-D2A35B0C8D67} = {D009BA72-08B7-4F6A-987D-806FFB01AE88}
 		{BF33FC29-2FBD-4DBC-97D7-9CF207130908} = {D009BA72-08B7-4F6A-987D-806FFB01AE88}
+		{82CAAF28-9AE0-4515-A47E-E436D0D3FA05} = {BC3238CA-1D20-4D33-B633-033A2627B1DF}
+		{DB50A83E-B1BA-44C1-BC55-760DCA114B2F} = {BC3238CA-1D20-4D33-B633-033A2627B1DF}
 	EndGlobalSection
 EndGlobal
