+++ build.bat
@@ -1,0 +1,4 @@
+@echo off
+
+dotnet tool restore
+dotnet run --project build/Build.fsproj -- %*
