+++ build.sh
@@ -1,0 +1,4 @@
+#!/bin/sh -x
+
+dotnet tool restore
+dotnet run --project build/Build.fsproj -- $@
