+++ tests/Thoth.Json.Tests.JavaScript/Thoth.Json.Tests.JavaScript.fsproj
@@ -1,0 +1,18 @@
+<Project Sdk="Microsoft.NET.Sdk">
+  <PropertyGroup>
+    <TargetFramework>netstandard2.0</TargetFramework>
+    <DefineConstants>$(DefineConstants);FABLE_COMPILER;FABLE_COMPILER_JAVASCRIPT</DefineConstants>
+  </PropertyGroup>
+  <ItemGroup>
+    <Compile Include="Main.fs" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\..\packages\Thoth.Json.Core\Thoth.Json.Core.fsproj" />
+    <ProjectReference Include="..\..\packages\Thoth.Json.JavaScript\Thoth.Json.JavaScript.fsproj" />
+    <ProjectReference Include="..\Thoth.Json.Tests\Thoth.Json.Tests.fsproj" />
+  </ItemGroup>
+  <ItemGroup>
+    <PackageReference Include="Fable.Core" />
+    <PackageReference Include="Fable.Mocha" />
+  </ItemGroup>
+</Project>
